monitoring_bot/requirements.txt
python-dotenv
python-json-logger
colorlog
python-telegram-bot
python-telegram-bot[job-queue]
requests
